--- src/models/SuperappData.ts
@@ -1,6 +1,0 @@
-export interface SuperappAuthItem {
-	service_id: number;
-	current_scopes?: string[];
-	token?: string;
-	expires?: string;
-}
